# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StockWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary Search for stocks by name or symbol
 */
export const SearchStocksQueryParams = zod.object({
  q: zod.coerce.string(),
});

export const SearchStocksResponseItem = zod.object({
  symbol: zod.string(),
  description: zod.string(),
  type: zod.string(),
  displaySymbol: zod.string(),
});
export const SearchStocksResponse = zod.array(SearchStocksResponseItem);

/**
 * @summary Get current stock quote
 */
export const GetStockQuoteParams = zod.object({
  symbol: zod.coerce.string(),
});

export const GetStockQuoteResponse = zod.object({
  symbol: zod.string(),
  c: zod.number().describe("Current price"),
  d: zod.number().describe("Change"),
  dp: zod.number().describe("Percent change"),
  h: zod.number().describe("High price of the day"),
  l: zod.number().describe("Low price of the day"),
  o: zod.number().describe("Open price of the day"),
  pc: zod.number().describe("Previous close price"),
  t: zod.number().describe("Unix timestamp"),
});

/**
 * @summary Get historical price data (OHLCV)
 */
export const GetStockCandlesParams = zod.object({
  symbol: zod.coerce.string(),
});

export const getStockCandlesQueryResolutionDefault = `D`;

export const GetStockCandlesQueryParams = zod.object({
  resolution: zod
    .union([
      zod.literal(1),
      zod.literal(5),
      zod.literal(15),
      zod.literal(30),
      zod.literal(60),
      zod.literal("D"),
      zod.literal("W"),
      zod.literal("M"),
    ])
    .default(getStockCandlesQueryResolutionDefault),
  from: zod.coerce.number().optional(),
  to: zod.coerce.number().optional(),
});

export const GetStockCandlesResponse = zod.object({
  c: zod.array(zod.number()).describe("Close prices"),
  h: zod.array(zod.number()).describe("High prices"),
  l: zod.array(zod.number()).describe("Low prices"),
  o: zod.array(zod.number()).describe("Open prices"),
  v: zod.array(zod.number()).describe("Volumes"),
  t: zod.array(zod.number()).describe("Unix timestamps"),
  s: zod.string().describe("Status (ok or no_data)"),
});

/**
 * @summary Get company profile and basic info
 */
export const GetCompanyProfileParams = zod.object({
  symbol: zod.coerce.string(),
});

export const GetCompanyProfileResponse = zod.object({
  name: zod.string(),
  ticker: zod.string(),
  country: zod.string().nullable(),
  currency: zod.string().nullable(),
  exchange: zod.string().nullable(),
  ipo: zod.string().nullable(),
  marketCapitalization: zod.number().nullable(),
  shareOutstanding: zod.number().nullable(),
  logo: zod.string().nullable(),
  finnhubIndustry: zod.string().nullable(),
  weburl: zod.string().nullish(),
  phone: zod.string().nullish(),
});

/**
 * @summary Get technical indicators (RSI, MACD, MA)
 */
export const GetStockTechnicalsParams = zod.object({
  symbol: zod.coerce.string(),
});

export const getStockTechnicalsQueryResolutionDefault = `D`;

export const GetStockTechnicalsQueryParams = zod.object({
  resolution: zod.coerce
    .string()
    .default(getStockTechnicalsQueryResolutionDefault),
});

export const GetStockTechnicalsResponse = zod.object({
  symbol: zod.string(),
  rsi: zod.object({
    values: zod.array(zod.number().nullable()),
    timestamps: zod.array(zod.number()),
  }),
  macd: zod.object({
    macdLine: zod.array(zod.number().nullable()),
    signalLine: zod.array(zod.number().nullable()),
    histogram: zod.array(zod.number().nullable()),
    timestamps: zod.array(zod.number()),
  }),
  sma20: zod.object({
    values: zod.array(zod.number().nullable()),
    timestamps: zod.array(zod.number()),
  }),
  sma50: zod.object({
    values: zod.array(zod.number().nullable()),
    timestamps: zod.array(zod.number()),
  }),
  sma200: zod.object({
    values: zod.array(zod.number().nullable()),
    timestamps: zod.array(zod.number()),
  }),
});

/**
 * @summary Compare multiple stocks performance
 */
export const compareStocksQueryResolutionDefault = `D`;

export const CompareStocksQueryParams = zod.object({
  symbols: zod.coerce.string().describe("Comma-separated list of symbols"),
  resolution: zod.coerce.string().default(compareStocksQueryResolutionDefault),
});

export const CompareStocksResponseItem = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  candles: zod.object({
    c: zod.array(zod.number()).describe("Close prices"),
    h: zod.array(zod.number()).describe("High prices"),
    l: zod.array(zod.number()).describe("Low prices"),
    o: zod.array(zod.number()).describe("Open prices"),
    v: zod.array(zod.number()).describe("Volumes"),
    t: zod.array(zod.number()).describe("Unix timestamps"),
    s: zod.string().describe("Status (ok or no_data)"),
  }),
  currentPrice: zod.number(),
  changePercent: zod.number(),
});
export const CompareStocksResponse = zod.array(CompareStocksResponseItem);

/**
 * @summary Get current user's watchlist
 */
export const GetWatchlistResponseItem = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  addedAt: zod.coerce.date(),
});
export const GetWatchlistResponse = zod.array(GetWatchlistResponseItem);

/**
 * @summary Add a stock to watchlist
 */
export const addToWatchlistBodySymbolMax = 20;

export const AddToWatchlistBody = zod.object({
  symbol: zod.string().min(1).max(addToWatchlistBodySymbolMax),
});

/**
 * @summary Remove a stock from watchlist
 */
export const RemoveFromWatchlistParams = zod.object({
  symbol: zod.coerce.string(),
});

/**
 * @summary Get live prices for all watchlist items
 */
export const GetWatchlistPricesResponseItem = zod.object({
  symbol: zod.string(),
  currentPrice: zod.number(),
  changePercent: zod.number(),
  change: zod.number(),
});
export const GetWatchlistPricesResponse = zod.array(
  GetWatchlistPricesResponseItem,
);

/**
 * @summary Send a message to the AI investment advisor
 */
export const SendAiChatMessageBody = zod.object({
  messages: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
  symbol: zod.string().nullish().describe("Optional stock symbol for context"),
});

export const SendAiChatMessageResponse = zod.object({
  message: zod.string(),
});