# lib/api-zod/src/generated/types/companyProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StockWatch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyProfile {
  name: string;
  ticker: string;
  /** @nullable */
  country: string | null;
  /** @nullable */
  currency: string | null;
  /** @nullable */
  exchange: string | null;
  /** @nullable */
  ipo: string | null;
  /** @nullable */
  marketCapitalization: number | null;
  /** @nullable */
  shareOutstanding: number | null;
  /** @nullable */
  logo: string | null;
  /** @nullable */
  finnhubIndustry: string | null;
  /** @nullable */
  weburl?: string | null;
  /** @nullable */
  phone?: string | null;
}