# lib/api-zod/src/generated/types/stockComparisonItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StockWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StockCandles } from "./stockCandles";

export interface StockComparisonItem {
  symbol: string;
  name: string;
  candles: StockCandles;
  currentPrice: number;
  changePercent: number;
}