# lib/api-zod/src/generated/types/technicalIndicators.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StockWatch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TechnicalIndicatorsMacd } from "./technicalIndicatorsMacd";
import type { TechnicalIndicatorsRsi } from "./technicalIndicatorsRsi";
import type { TechnicalIndicatorsSma20 } from "./technicalIndicatorsSma20";
import type { TechnicalIndicatorsSma200 } from "./technicalIndicatorsSma200";
import type { TechnicalIndicatorsSma50 } from "./technicalIndicatorsSma50";

export interface TechnicalIndicators {
  symbol: string;
  rsi: TechnicalIndicatorsRsi;
  macd: TechnicalIndicatorsMacd;
  sma20: TechnicalIndicatorsSma20;
  sma50: TechnicalIndicatorsSma50;
  sma200: TechnicalIndicatorsSma200;
}